Covid.jl

Covid.CSV_to_DataFrame — Function

CSV_to_DataFrame(path_url::String, encoding::String="UTF-8")::DataFrame

Lee un archivo CSV con el encoding indicado y regresa un DataFrame.

Ejemplo

julia> df = DataFrameEncode("datos.csv")
julia> df_latin1 = DataFrameEncode("datos.csv", "LATIN1")

Los encodings soportados dependen de la plataforma, obtén la lista de la siguiente manera.

julia> using StringEncodings
julia> encodings()

Covid.Select — Method

Select(Tabla::DataFrame,query::Vector{String})::DataFrame

Selecciona una o varias columnas del DataFrame puede ser por nombre o por número de columna y regresa un nuevo DataFrame con las columnas seleccionadas.

Ejemplo

julia> tabla = DataFrame(A = 1:3, B= 1:3)
3×2 DataFrame
 Row │ A      B
     │ Int32  Int32
─────┼──────────────
   1 │     1      1
   2 │     2      2
   3 │     3      3

julia> q1 = Selects.Select(tabla,["1"])
3×1 DataFrame
 Row │ A
     │ Int32
─────┼───────
   1 │     1
   2 │     2
   3 │     3

Covid.data_check — Function

data_check(path_url::String, type::String="PATH", encoding::String="UTF-8")::DataFrame

Crea un DataFrame dado un archivo CSV o una liga al archivo. Se pude especificar el encoding.

Ejemplo

julia> url = "http://www.conapo.gob.mx/work/models/OMI/Datos_Abiertos/DA_IAIM/IAIM_Municipio_2010.csv"
julia> first(data_check(url, "URL", "LATIN1"))
julia> first(data_check("prueba.csv"))

Covid.fechahoy — Method

fechahoy()::String

Crea un string con la fecha de hoy utilizando el formato "yyyymmdd". Año con cuarto dígitos, mes y día con dos.

Ejemplo

julia> fechahoy()
"20210112"

Covid.jsonparse — Method

jsonparse(url::String)::Dict

Hace un http request al url especificado y convierte el json obtenido del sitio web en un diccionario. En caso de que el servidor devuelva un status distinto a 200, se arroja un error.

Ejemplo

julia> datos = jsonparse("https://sitioweb.com/datos.json")

Covid.poblacion_entidad — Method

poblacion_entidad(token_INEGI::String, cve_entidad::String)::poblacion

Regresa un una estructura poblacion con los datos más recientes, por entidad federativa, proporcionados por la API de Indicadores del INEGI. Requiere el token (token_INEGI) de la API, puede obtenerse aquí. La estructura poblacion contiene los siguientes datos.

lugar

población total

población total hombres

población total mujeres

porcentaje de hombres

porcentaje de mujeres

porcentaje de población que se considera indígena

Note

Área geoestadística estatal (AGEE)

La entidad federativa se codifica de acuerdo con el orden alfabético de sus nombres oficiales, con una longitud de dos dígitos, a partir del 01 en adelante, según el número de entidades federativas que dispongan las leyes vigentes; en este momento son 32 entidades federativas (Aguascalientes 01, Baja California 02,... y Zacatecas 32). Las puedes consultar aquí.

Clave Entidad	Entidad

01	Aguascalientes

02	Baja California

03	Baja California Sur

⋮	⋮

30	Veracruz de Ignacio de la Llave

31	Yucatán

32	Zacatecas

Ejemplo

julia> struct poblacion
  lugar::String
  total::Float64
  hombres::Float64
  mujeres::Float64
  porcentaje_hombres::Float64
  porcentaje_mujeres::Float64
  porcentaje_indigena::Float64
end

julia> popu = poblacion_entidad(token, "31")
Yucatán

Población total 2102259.00 (53.06 hab/km²)
Hombres 963333.00 (49.00%)
Mujeres 992244.00 (51.00%)
Porcentaje que se considera indígena 65.40%

julia> popu.porcentaje_indigena
65.403459

Covid.poblacion_mexico — Method

poblacion_mexico(token_INEGI::String)::poblacion

Regresa un una estructura poblacion con los datos más recientes, a nivel nacional, proporcionados por la API de Indicadores del INEGI. Requiere el token (token_INEGI) de la API, puede obtenerse aquí.

Ejemplo

julia> popu = poblacion_mexico(token)
México

Población total 119938473.00 (60.96 hab/km²)
Hombres 54855231.00 (48.57%)
Mujeres 57481307.00 (51.43%)
Porcentaje que se considera indígena 21.50%

Covid.poblacion_municipio — Method

poblacion_municipio(token_INEGI::String, cve_entidad::String, cve_municipio::String)::poblacion

Regresa un una estructura poblacion con los datos más recientes, por municipio, proporcionados por la API de Indicadores del INEGI. Requiere el token (token_INEGI) de la API, puede obtenerse aquí.

Note

Área geoestadística municipal (AGEM)

La clave del municipio está formada por tres números que se asignan de manera ascendente a partir del 001, de acuerdo con el orden alfabético de los nombres de los municipios, aunque a los creados posteriormente a la clavificación inicial, se les asigna la clave geoestadística conforme se vayan creando. Las puedes consultar aquí.

Clave Entidad	Nombre Entidad	Clave Municipio	Nombre Municipio

01	Aguascalientes	001	Aguascalientes

01	Aguascalientes	002	Asientos

01	Aguascalientes	003	Calvillo

⋮	⋮	⋮	⋮

32	Zacatecas	056	Zacatecas

32	Zacatecas	057	Trancoso

32	Zacatecas	058	Santa María de la Paz

Ejemplo

julia> popu = poblacion_municipio(token, "01", "002")
Aguascalientes, Asientos

Población total 45492.00 (84.63 hab/km²)
Hombres 22512.00 (48.95%)
Mujeres 22980.00 (51.05%)
Porcentaje que se considera indígena 3.64%

Covid.sumacolumna — Method

sumacolumna(tabla::DataFrame, col::Int)::Number
sumacolumna(tabla::DataFrame, col::String)::Number

Suma todos los valores de una determinada columna en un DataFrame. Para hacer referencia a que columna se desea sumar se pude usar la posición de la columna o el nombre que tiene.

Ejemplo

julia> df = data_check("datos.csv")
4×2 DataFrame
 Row │ x      y
     │ Int64  Int64
─────┼──────────────
   1 │     0     11
   2 │     2     12
   3 │     0     13
   4 │    40     14

julia> sumacolumna(df, 1)
42

julia> sumacolumna(df, "x")
42

Covid.sumafila — Method

sumafila(tabla::DataFrame, fila::Int)::Number

Suma todos los valores de una determinada fila en un DataFrame. La fila se especifica con la posición en la que se encuentra.

Ejemplo

julia> df = data_check("datos.csv")
4×2 DataFrame
 Row │ x      y
     │ Int64  Int64
─────┼──────────────
   1 │     0     11
   2 │     2     12
   3 │     0     13
   4 │    40     14

julia> sumafila(df, 2)
14

julia> sumafila(df, 4)
54

Covid.unzip — Function

unzip(path::String, dest::String="")

Descomprime y guarda el archivo en el destino indicado(dest), si no se proporciona un destino, se guarda en el directorio actual.

Ejemplo

julia> unzip("datos.zip")
julia> unzip("datos.zip", pwd()*"/datos")