Reporstat.jl

Paquete que tiene como objetivo agilizar la consulta de información poblacional de instituciones como INEGI, CONAPO y CONEVAL segregada por municipios para que, en conjunto con datos abiertos de cualquier índole, se facilite la realización de análisis y reportes estadísticos al usuario.

Indice

Reporstat.jl

Indice

Funciones

Población

Indicadores de Pobreza

Utilidades

Funciones

Reporstat.cargar_csv

Reporstat.clave

Reporstat.codigos_postales

Reporstat.codigos_postales_todos

Reporstat.contar_renglones

Reporstat.csv_a_DataFrame

Reporstat.edad_entidades

Reporstat.edad_municipios

Reporstat.edad_municipios_todos

Reporstat.fechahoy

Reporstat.filtrar

Reporstat.geografia

Reporstat.geografia_todos_municipios

Reporstat.idh

Reporstat.idh_todos_municipios

Reporstat.indicadores_pobreza

Reporstat.indicadores_pobreza_porcentaje

Reporstat.indicadores_pobreza_porcentaje_todos

Reporstat.indicadores_pobreza_todos

Reporstat.int_migratoria

Reporstat.int_migratoria_todos

Reporstat.jsonparse

Reporstat.poblacion_entidad

Reporstat.poblacion_mexico

Reporstat.poblacion_municipio

Reporstat.poblacion_todas_entidades

Reporstat.poblacion_todos_municipios

Reporstat.seleccionar

Reporstat.similitud_entidad

Reporstat.similitud_municipio

Reporstat.similitud_region

Reporstat.sumar_columna

Reporstat.sumar_fila

Reporstat.tasas_vitales

Reporstat.unzip

Población

Funciones que manejan datos proporcionados por INEGI, CONAPO y CONEVAL y proporcionan consultas por municipio. Detalles de año de actualización, consideraciones, y ejemplos proporcionados en cada apartado por función.

Reporstat.idh — Function

idh(cve_entidad::String, cve_municipio::String="")::DataFrame

Regresa el índice de desarrollo humano, los años promedio de escolaridad, los años esperados de escolaridad y los ingresos per cápita de una entidad o de un municipio en formato DataFrame, se debe especificar la clave para ambos parámetros, si solo se manda el parámetro cveentidad_ se regresará el IDH de la entidad. Los datos son obtenidos de la página oficial de las Naciones Unidas, se pueden consultar aquí.

Ejemplo

julia> idh(clave("Campeche"),"003").idh*100
1-element Array{Float64,1}:
 77.50874

julia> idh(clave("Campeche"),"003").idh[1]*100
77.50874

julia> idh(clave("Campeche"),"003")
1×8 DataFrame
 Row │ ent  mun  entidad   municipio  idh       anio_promedio_escolaridad  anios_esperados_escolaridad  ingreso_per_capita 
     │ Any  Any  Any       Any        Any       Any                        Any                          Any                
─────┼─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ 04   003  Campeche  Carmen     0.775087  9.19528                    12.6746                      18552.8

julia> idh(clave("Campeche"))
1×3 DataFrame
 Row │ ent  ent_nombre  idh  
     │ Any  Any         Any  
─────┼───────────────────────
   1 │ 04   Campeche    0.82

Reporstat.idh_todos_municipios — Function

idh_todos_municipios()::DataFrame

Regresa un DataFrame con todos los valores agregados del índice de desarrollo humano de todos los municipios. Los datos son obtenidos de la página oficial de las Naciones Unidas, se pueden consultar aquí.

Ejemplo

julia> idh_todos_municipios()
2456×8 DataFrame
  Row │ ent     mun     entidad              municipio                     idh       anio_promedio_escolaridad  anios_esperados_escolaridad  ingreso_per_capita 
      │ String  String  String               String                        Float64   Float64                    Float64                      Float64            
──────┼─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01      001     Aguascalientes       Aguascalientes                0.789432                    9.55307                      12.8703            17848.3
    2 │ 01      002     Aguascalientes       Asientos                      0.6615                      6.22975                      11.1364             6877.03
    3 │ 01      003     Aguascalientes       Calvillo                      0.669557                    5.95726                      10.8926             8764.94
  ⋮   │   ⋮       ⋮              ⋮                        ⋮                   ⋮                  ⋮                           ⋮                       ⋮
 2455 │ 32      057     Zacatecas            Trancoso                      0.673578                    6.24331                      10.8909             7555.7
 2456 │ 32      058     Zacatecas            Santa María de la Paz         0.668754                    5.86283                      12.143              8092.09

Reporstat.int_migratoria — Function

int_migratoria(cve_entidad::String, cve_municipio::String ="")::Float64

Devuelve el índice de intensidad migratoria del municipio indicado. En caso de omitir el parámetro cve_municipio, devuelve los datos de la entidad indicada. Datos obtenidos de aquí.

Ejemplo

julia> int_migratoria(clave("Campeche"),"003")
0.288

julia> int_migratoria(clave("Campeche"))
0.64

Reporstat.int_migratoria_todos — Function

int_migratoria_todos()::DataFrame

Regresa un DataFrame con los índices de intensidad migratoria de todos los municipios. Datos obtenidos de aquí.

Ejemplo

julia> int_migratoria_todos()
2469×5 DataFrame
  Row │ ent     ent_nombre      mun     mun_nom                       iaim     ⋯
      │ String  String          String  String                        Float64? ⋯
──────┼─────────────────────────────────────────────────────────────────────────
    1 │ 01      Aguascalientes  001     Aguascalientes                   1.92  ⋯
    2 │ 01      Aguascalientes  002     Asientos                         5.18
  ⋮   │   ⋮           ⋮           ⋮                  ⋮                   ⋮     ⋱
 2468 │ 32      Zacatecas       057     Trancoso                         4.141 ⋯
 2469 │ 32      Zacatecas       058     Santa María de la Paz           10.074

Reporstat.edad_entidades — Function

edad_entidades()::DataFrame

Da a conocer el primer y tercer cuartil, así como mediana(segundo cuartil) de las edades por entidad en formato DataFrame. Dichos datos de edades actualizados al año 2020 se obtuvieron de la página INEGI.

Ejemplo

julia> edad_entidades()
32×5 DataFrame
  Row │ entidad  entidad_nombre       Q1             Q2            Q3
      │ String   String                 Float64        Float64       Float64
──────┼───────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes            13           26              43
    2 │ 02       Baja California           15           28              44
    3 │ 03       Baja California Sur       14           28              43
   ⋮        ⋮             ⋮              ⋮             ⋮                ⋮

Reporstat.edad_municipios — Function

edad_municipios(cve_entidad::String,cve_municipio::String)::DataFrame

Da a conocer el primer y tercer cuartil, así como mediana(segundo cuartil) de las edades del municipio indicado en formato DataFrame. Dichos datos de edades actualizados al año 2020 se obtuvieron de la página INEGI.

Ejemplo

julia> edad_municipios("01", "001")
1×7 DataFrame
  Row │ entidad  entidad_nombre  municipio  municipio_nombre    Q1             Q2            Q3
      │ String   String          String     String            Float64        Float64       Float64
──────┼──────────────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes  001        Aguascalientes        14           28              46

Reporstat.edad_municipios_todos — Function

edad_municipios_todos()::DataFrame

Da a conocer el primer y tercer cuartil, así como mediana(segundo cuartil) de las edades, a nivel nacional segregado por municipio en formato DataFrame. Dichos datos de edades actualizados al año 2020 se obtuvieron de la página INEGI.

Ejemplo

julia> edad_municipios_todos()
2469×7 DataFrame
  Row │ entidad  entidad_nombre  municipio  municipio_nombre    Q1             Q2            Q3
      │ String   String          String     String            Float64        Float64       Float64
──────┼──────────────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes  001        Aguascalientes        14           28              46
    2 │ 01       Aguascalientes  002        Asientos              11           24              42
    3 │ 01       Aguascalientes  003        Calvillo              12           27              46
   ⋮           ⋮             ⋮       ⋮             ⋮                   ⋮            ⋮               ⋮

Reporstat.tasas_vitales — Function

tasas_vitales(cve_entidad::String, cve_municipio::String="")::DataFrame

Proporciona un DataFrame con las tasas de natalidad, fecundidad y mortalidad del municipio indicado. En caso de omitir el parámetro cve_municipio, se mostrarán datos de la entidad indicada. Datos obtenidos del registro de nacimientos (2019), defunciones generales (2019) y población de mujeres en edad fértil (15-45 años, 2020) del INEGI.

Ejemplo

julia> tasas_vitales("01", "001")
1×3 DataFrame
 Row │ Natalidad  Fecundidad  Mortalidad
     │ Float64    Float64     Float64
─────┼───────────────────────────────────
   1 │ 0.0430915   0.0915221   0.0221098

Reporstat.geografia — Function

geografia(cve_entidad::String,cve_municipio::String ="")::DataFrame

Devuelve un DataFrame con la latitud, longitud y altitud promedio del municipio indicado. En caso de omitir el parámetro cve_municipio, devuelve los datos de la entidad indicada. Datos obtenidos de aquí.

Se pueden hacer consultas de una entidad o de un municipio.

julia> geografia(clave("Oaxaca"),"003")
1×5 DataFrame
 Row │ ent  mun  latitud       longitud       altitud
     │ Any  Any  Any           Any            Any
─────┼────────────────────────────────────────────────
   1 │ 20   003  17°04´10.549  095°58´04.929  1486

julia> geografia(clave("Oaxaca"),"003").latitud
1-element Array{Any,1}:
 "17°04´10.549"

julia> geografia(clave("Oaxaca"),"003").altitud
1-element Array{Any,1}:
 "1486"

julia> geografia(clave("Campeche"))
1×4 DataFrame
 Row │ ent  latitud       longitud       altitud
     │ Any  Any           Any            Any
─────┼───────────────────────────────────────────
   1 │ 04   19°01´17.138  092°27´54.859  14

Reporstat.geografia_todos_municipios — Function

geografia_todos_municipios()::DataFrame
Devuelve los datos geográficos de todos los municipios(latitud, longitud y altitud promedio), puedes consultar la información [aquí](https://www.inegi.org.mx/app/ageeml/#).

Ejemplo

julia> geografia_todos_municipios()
2469×7 DataFrame
  Row │ ent     nom_ent         mun     nom_mun                       latitud  ⋯
      │ String  String          String  String                        String   ⋯
──────┼─────────────────────────────────────────────────────────────────────────
    1 │ 01      Aguascalientes  001     Aguascalientes                22°03´26 ⋯
    2 │ 01      Aguascalientes  002     Asientos                      22°17´45
  ⋮   │   ⋮           ⋮           ⋮                  ⋮                     ⋮   ⋱
 2468 │ 32      Zacatecas       057     Trancoso                      22°49´06 ⋯
 2469 │ 32      Zacatecas       058     Santa María de la Paz         21°33´55

Reporstat.codigos_postales — Function

codigos_postales()::DataFrame

Proporciona los códigos postales del municpio indicado, en un DataFrame. Los datos son obtenidos del Servicio Postal Mexicano.

Ejemplo

julia> codigos_postales("01", "001")
1×6 DataFrame
  Row │ entidad  entidad_nombre  municipio  municipio_nombre  número de códigos postales  códigos postales
      │ String   String          String     String            Int64                       String
──────┼────────────────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes  001        Aguascalientes                           599  20000;20010;20010 ⋯

Reporstat.codigos_postales_todos — Function

codigos_postales_todos()::DataFrame

Proporciona todos los códigos postales de México, segregados por municipio, en un DataFrame. Los datos son obtenidos del Servicio Postal Mexicano.

Ejemplo

julia> codigos_postales_todos()
2465×6 DataFrame
  Row │ entidad  entidad_nombre  municipio  municipio_nombre  número de códigos postales  códigos postales
      │ String   String          String     String            Int64                       String
──────┼────────────────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes  001        Aguascalientes                           599  20000;20010;20010 ⋯
    2 │ 01       Aguascalientes  002        Asientos                                  82  20700;20700;20700 ⋯
  ⋮   │    ⋮           ⋮             ⋮                   ⋮                ⋮                               ⋮

Reporstat.poblacion_mexico — Function

poblacion_mexico(token_INEGI::String="")::DataFrame

Regresa un DataFrame con los datos más recientes, a nivel nacional, proporcionados por la API de Indicadores del INEGI. Requiere el token (token_INEGI) de la API, puede obtenerse aquí. Se puede proporcionar el token directamente o por medio de una variable de entorno llamada de la misma manera, token_INEGI.

Ejemplo

julia> ENV["token_INEGI"] = "00000000-0000-0000-0000-000000000000"
"00000000-0000-0000-0000-000000000000"

julia> poblacion_mexico()
1×9 DataFrame
 Row │ lugar   total      hombres   mujeres   porcentaje_hombres  porcentaje_mujeres  porcentaje_indigena  densidad_poblacion  extesion_territorial
     │ String  Int64      Int64     Int64     Float64             Float64             Float64              Float64             Float64
─────┼──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ México  126014024  61473390  64540634              48.783              51.217              21.4965             64.2717             1.96065e6

Reporstat.poblacion_entidad — Function

poblacion_entidad(cve_entidad::String, token_INEGI::String="")::poblacion

Regresa un una DataFrame con los datos más recientes, por entidad federativa, proporcionados por la API de Indicadores del INEGI. Requiere el token (token_INEGI) de la API, puede obtenerse aquí. Se puede proporcionar el token directamente o por medio de una variable de entorno, de la siguiente manera.

julia> ENV["token_INEGI"] = "00000000-0000-0000-0000-000000000000"

El DataFrame resultante contiene los siguientes datos.

lugar

población total

densidad de población (habitantes por kilómetro cuadrado)

extensión terrotorial (kilómetros cuadrados)

población total hombres

población total mujeres

porcentaje de hombres

porcentaje de mujeres

porcentaje de población que se considera indígena

Note

Área geoestadística estatal (AGEE)

La entidad federativa se codifica de acuerdo con el orden alfabético de sus nombres oficiales, con una longitud de dos dígitos, a partir del 01 en adelante, según el número de entidades federativas que dispongan las leyes vigentes; en este momento son 32 entidades federativas (Aguascalientes 01, Baja California 02,... y Zacatecas 32). Las puedes consultar aquí.

Clave Entidad	Entidad

01	Aguascalientes

02	Baja California

03	Baja California Sur

⋮	⋮

30	Veracruz de Ignacio de la Llave

31	Yucatán

32	Zacatecas

Ejemplo

julia> popu = poblacion_entidad("31", token)
1×9 DataFrame
 Row │ lugar    total    hombres  mujeres  porcentaje_hombres  porcentaje_mujeres  porcentaje_indigena  densidad_poblacion  extesion_territorial
     │ String   Int64    Int64    Int64    Float64             Float64             Float64              Float64             Float64
─────┼───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ Yucatán  2320898  1140279  1180619             49.1309             50.8691              65.4035             58.7206               39524.4

Reporstat.poblacion_municipio — Function

poblacion_municipio(cve_entidad::String, cve_municipio::String, token_INEGI::String="")::DataFrame

Regresa un DataFrame con los datos más recientes, por municipio, proporcionados por la API de Indicadores del INEGI. Requiere el token (token_INEGI) de la API, puede obtenerse aquí. Se puede proporcionar el token directamente o por medio de una variable de entorno llamada de la misma manera, token_INEGI.

Note

Área geoestadística municipal (AGEM)

La clave del municipio está formada por tres números que se asignan de manera ascendente a partir del 001, de acuerdo con el orden alfabético de los nombres de los municipios, aunque a los creados posteriormente a la clavificación inicial, se les asigna la clave geoestadística conforme se vayan creando. Las puedes consultar aquí.

Clave Entidad	Nombre Entidad	Clave Municipio	Nombre Municipio

01	Aguascalientes	001	Aguascalientes

01	Aguascalientes	002	Asientos

01	Aguascalientes	003	Calvillo

⋮	⋮	⋮	⋮

32	Zacatecas	056	Zacatecas

32	Zacatecas	057	Trancoso

32	Zacatecas	058	Santa María de la Paz

Ejemplo

julia> ENV["token_INEGI"] = "00000000-0000-0000-0000-000000000000"
"00000000-0000-0000-0000-000000000000"

julia> popu = poblacion_municipio("01", "002")
1×9 DataFrame
 Row │ lugar                     total  hombres  mujeres  porcentaje_hombres  porcentaje_mujeres  porcentaje_indigena  densidad_poblacion  extesion_territorial
     │ String                    Int64  Int64    Int64    Float64             Float64             Float64              Float64             Float64
─────┼──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ Aguascalientes, Asientos  51536    25261    26275             49.0162             50.9838              3.63938             93.8709               549.009

Reporstat.poblacion_todas_entidades — Function

poblacion_todas_entidades()::DataFrame

Regresa un DataFrame con los datos poblacionales de todas las entidades.

clave de entidad

nombre oficial de la entidad

población total

densidad de población (habitantes por kilómetro cuadrado)

extensión territorial (kilómetros cuadrados)

población total hombres

población total mujeres

porcentaje de hombres

porcentaje de mujeres

porcentaje de población que se considera indígena

Ejemplo

julia> poblacion_todas_entidades()
32×10 DataFrame
 Row │ entidad  entidad_nombre     total     densidad   extension_territorial  hombres    mujeres    ⋯
     │ String   String             Int64     Float64    Float64                Int64      Int64      ⋯
─────┼───────────────────────────────────────────────────────────────────────────────────────────────
   1 │ 01       Aguascalientes      1425607   253.862            5615.67        69668300   72892400  ⋯
   2 │ 02       Baja California     3769020    52.7505          71450.0        190058900  186843100        
   ⋮ │ ⋮             ⋮                 ⋮            ⋮           ⋮                 ⋮                      ⋱
  32 │ 32       Zacatecas           1622138    21.5494          75275.3         79105800   83108000  ⋯

Reporstat.poblacion_todos_municipios — Function

poblacion_todos_municipios()::DataFrame

Regresa un DataFrame con los datos poblacionales de todos los municipios.

nombre del lugar

clave de entidad

nombre de la entidad

clave de municipio

nombre de municipio

población total

densidad de población (habitantes por kilómetro cuadrado)

extensión territorial (kilómetros cuadrados)

población total hombres

población total mujeres

porcentaje de hombres

porcentaje de mujeres

porcentaje de población que se considera indígena

Ejemplo

julia> poblacion_todos_municipios()
2469×11 DataFrame
  Row │ entidad  entidad_nombre  municipio  municipio_nombre              total   densidad   hombres   mujeres   porcentaje_hombres  porcentajes_mujeres ⋯
      │ String   String          String     String                        Int64   Float64    Int64     Int64     Float64             Float64             ⋯
──────┼──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes  001        Aguascalientes                797010   744.58     386429   410581.0             48.5335              51.4665 ⋯
    2 │ 01       Aguascalientes  002        Asientos                      797010   744.58     386429   410581.0             48.5335              51.4665
    ⋮   │    ⋮           ⋮             ⋮                   ⋮                   ⋮           ⋮         ⋮         ⋮              ⋮                    ⋮          ⋱
 2468 │ 32       Zacatecas       057        Trancoso                      138176   331.026     66297    71879.0             48.482               51.518  ⋯
 2469 │ 32       Zacatecas       058        Santa María de la Paz          16934    87.9192     8358     8576.0             48.962               51.038

Reporstat.similitud_region — Function

similitud_region(region::String)::Array{Array{String}}

Proporciona dos arreglos dentro de un arreglo. El primer arreglo contiene todas las entidades similares a region y el segundo todos los municipios similares a region.

Ejemplo

julia> similitud_region("jalisto")
2-element Array{Array{String,N} where N,1}:
 ["Jalisco"]
 ["Xalisco", "Naolinco", "Xaloztoc", "Calvillo", "Saltillo"]

Reporstat.similitud_entidad — Function

similitud_entidad(entidad::String)::Array{String}

Proporciona un arreglo con todas las entidades con un nombre igual o similar a entidad.

Ejemplo

julia> similitud_entidad("oajaca")
1-element Array{String,1}:
 "Oaxaca"

julia> clave(similitud_entidad("oaxjaca")[1])
"29"

Reporstat.similitud_municipio — Function

similitud_municipio(municipio::String)::Array{String}

Proporciona un arreglo con todos los municipios con un nombre igual o similar a municipio.

Ejemplo

julia> similitud_municipio("tequixciapn")
3-element Array{String,1}:
 "Tequisquiapan"
 "Atlequizayan"
 "Tequixquiac"

julia> clave(similitud_municipio("tequixciapn")[end])
"096"

Indicadores de Pobreza

Datos proporcionados por el Consejo Nacional de Evaluación de la Política de Desarrollo Social, actualizados al año 2015.

Diccionario de Datos, Indicadores de pobreza municipal (2015)

Abreviación	Significado

pobreza	Población en situación de pobreza

pobreza_e	Población en situación de pobreza extrema

pobreza_m	Población en situación de pobreza moderada

vul_car	Población vulnerable por carencias

vul_ing	Población vulnerable por ingreso

npnv	Población no pobre y no vulnerable

ic_rezedu	Población con rezago educativo

ic_asalud	Población con carencia por acceso a los servicios de salud

ic_segsoc	Población con carencia por acceso a la seguridad social

ic_cv	Población con carencia por calidad y espacios en la vivienda

ic_sbv	Población con carencia por acceso a los servicios básicos en la vivienda

ic_ali	Población con carencia por acceso a la alimentación

carencias	Población con al menos una carencia social

carencias3	Población con al menos tres carencias sociales

plb	Población con ingreso inferior a la línea de bienestar

plbm	Población con ingreso inferior a la línea de bienestar mínimo

Reporstat.indicadores_pobreza — Function

indicadores_pobreza(cve_entidad::String,cve_municipio::String)::DataFrame

Proporciona el número de personas que cumple con los indicadores de pobreza según el CONEVAL, del municipio indicado.

Los datos son obtenidos de la página oficial de datos abiertos del gobierno federal de México datos.gob.mx Consulta el Diccionario de Datos, Indicadores de pobreza municipal (2015)

Ejemplo

julia> df = indicadores_pobreza("01", "001")
1×20 DataFrame
  Row │ entidad  entidad_nombre       municipio  municipio_nombre pobreza  pobreza_e  pobreza_m ⋯
      │ String   String               String     String           Int64    Int64      Int64     ⋯
──────┼────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes       001        Aguascalientes   224949      13650     211299 ⋯

Reporstat.indicadores_pobreza_porcentaje — Function

indicadores_pobreza_porcentaje(cve_entidad::String,cve_municipio::String)::DataFrame

Proporciona el porcentaje de personas que cumple con los indicadores de pobreza según el CONEVAL, del municipio indicado.

Los datos son obtenidos de la página oficial de datos abiertos del gobierno federal de México datos.gob.mx Consulta el Diccionario de Datos, Indicadores de pobreza municipal (2015)

julia> df = indicadores_pobreza_porcentaje("01", "001")
1×20 DataFrame
  Row │ entidad  entidad_nombre       municipio  municipio_nombre pobreza  pobreza_e  pobreza_m ⋯
      │ String   String               String     String           Float64  Float64    Float64   ⋯
──────┼─────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes       001        Aguascalientes      26.1        1.6       24.5 ⋯

Reporstat.indicadores_pobreza_todos — Function

indicadores_pobreza_todos()::DataFrame

Proporciona el número de personas que cumple con los indicadores de pobreza según el CONEVAL, a nivel nacional segregado por municipios.

Los datos son obtenidos de la página oficial de datos abiertos del gobierno federal de México datos.gob.mx Consulta el Diccionario de Datos, Indicadores de pobreza municipal (2015)

julia> df = indicadores_pobreza_todos()
2457×20 DataFrame
  Row │ entidad  entidad_nombre       municipio  municipio_nombre pobreza  pobreza_e  pobreza_m ⋯
      │ String   String               String     String           Int64    Int64      Int64     ⋯
──────┼────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes       001        Aguascalientes   224949      13650     211299 ⋯
    2 │ 01       Aguascalientes       002        Asientos          25169       2067      23101 ⋯
   ⋮  │ ⋮             ⋮               ⋮               ⋮            ⋮            ⋮          ⋮

Reporstat.indicadores_pobreza_porcentaje_todos — Function

indicadores_pobreza_porcentaje_todos()::DataFrame

Proporciona el porcentaje de personas que cumple con los indicadores de pobreza según el CONEVAL, a nivel nacional segregado por municipios.

Los datos son obtenidos de la página oficial de datos abiertos del gobierno federal de México datos.gob.mx Consulta el Diccionario de Datos, Indicadores de pobreza municipal (2015)

julia> df = indicadores_pobreza_porcentaje_todos()
2457×20 DataFrame
  Row │ entidad  entidad_nombre       municipio  municipio_nombre pobreza  pobreza_e  pobreza_m ⋯
      │ String   String               String     String           Float64  Float64    Float64   ⋯
──────┼─────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes       001        Aguascalientes      26.1        1.6       24.5 ⋯
    2 │ 01       Aguascalientes       002        Asientos            54.0        4.4       49.5 ⋯
   ⋮  │ ⋮             ⋮               ⋮               ⋮            ⋮            ⋮          ⋮

Utilidades

Utilidades varias que tienen como objetivo facilitar la manipulación de archivos CSV, DataFrames, entre otras.

Reporstat.clave — Function

clave(id::String)::String

Toma como parámetro el nombre de algún municipio o entidad y regresa la clave del mismo.

Ejemplo

julia> clave("Campeche")
"04"
julia> clave("Calakmul")
"010"

Reporstat.cargar_csv — Function

cargar_csv(path_url::String, type::String="PATH", encoding::String="UTF-8")::DataFrame

Crea un DataFrame dado un archivo CSV o una liga al archivo. Se pude especificar el encoding.

Ejemplo

julia> url = "http://www.conapo.gob.mx/work/models/OMI/Datos_Abiertos/DA_IAIM/IAIM_Municipio_2010.csv"
julia> first(cargar_csv(url, "URL", "LATIN1"))
julia> first(cargar_csv("prueba.csv"))

Reporstat.csv_a_DataFrame — Function

csv_a_DataFrame(path_url::String, encoding::String="UTF-8")::DataFrame

Lee un archivo CSV con el encoding indicado y regresa un DataFrame.

Ejemplo

julia> df = DataFrameEncode("datos.csv")
julia> df_latin1 = DataFrameEncode("datos.csv", "LATIN1")

Los encodings soportados dependen de la plataforma, obtén la lista de la siguiente manera.

julia> using StringEncodings
julia> encodings()

Reporstat.unzip — Function

unzip(path::String, dest::String="")

Descomprime y guarda el archivo en el destino indicado(dest), si no se proporciona un destino, se guarda en el directorio actual.

Ejemplo

julia> unzip("datos.zip")
julia> unzip("datos.zip", pwd()*"/datos")

Reporstat.jsonparse — Function

jsonparse(url::String)::Dict

Del url indicado,convierte el json obtenido del sitio web en un diccionario. En caso de que el servidor devuelva un status distinto a 200, se arroja un error.

Ejemplo

julia> datos = jsonparse("https://sitioweb.com/datos.json")

Reporstat.fechahoy — Function

fechahoy()::String

Crea un string con la fecha de hoy utilizando el formato "yyyymmdd". Año con cuarto dígitos, mes y día con dos.

Ejemplo

julia> fechahoy()
"20210112"

Reporstat.filtrar — Function

filtrar(tabla::DataFrame, condiciones...)::DataFrame

Filtra un DataFrame de acuerdo a las condiciones indicadas en los parametros (los parámetros actúan como expresiones). Los nombres de las columnas deben siempre tener el prefijo : , y puede ser especificada por nombre o por numero de columna.

Ejemplo

julia> dt= DataFrame(A = [1,2,2,3],B = ["A","A","B","BB"])
4×2 DataFrame
 Row │ A      B
     │ Int32  String
─────┼───────────────
   1 │     1  A
   2 │     2  A
   3 │     2  B
   4 │     3  BB

julia> filtrar(dt, ":1 == 2")
2×2 DataFrame
 Row │ A    B
     │ Any  Any
─────┼──────────
   1 │ 2    A
   2 │ 2    B

En caso de que se trate de una string se deben siempre encerrar entre comillas simples.

julia> filtrar(dt, "'A' == :2")
2×2 DataFrame
 Row │ A    B
     │ Any  Any
─────┼──────────
   1 │ 1    A
   2 │ 2    A

Las condiciones se deben poner en parámetros diferentes.

julia> filtrar(dt, "2 == :A" , "'A'== :B")
1×2 DataFrame
 Row │ A    B
     │ Any  Any
─────┼──────────
   1 │ 2    A

Reporstat.seleccionar — Function

seleccionar(Tabla::DataFrame, query::Vector{String})::DataFrame

Selecciona una o varias columnas del DataFrame puede ser por nombre o por número de columna y regresa un nuevo DataFrame con las columnas seleccionadas.

Ejemplo

julia> tabla = DataFrame(A = 1:3, B= 1:3)
3×2 DataFrame
 Row │ A      B
     │ Int32  Int32
─────┼──────────────
   1 │     1      1
   2 │     2      2
   3 │     3      3

julia> q1 = seleccionar(tabla,["1"])
3×1 DataFrame
 Row │ A
     │ Int32
─────┼───────
   1 │     1
   2 │     2
   3 │     3

Reporstat.contar_renglones — Function

contar_renglones(tabla::DataFrame, condiciones...)::Number

Llama internamente a la función Reporstat.filtrar con los mismos argumentos y regresa el número de renglones que tiene el DataFrame que retorna Reporstat.filtrar

Ejemplo

julia> dt
4×2 DataFrame
 Row │ A      B
     │ Int32  String
─────┼───────────────
   1 │     1  A
   2 │     2  A
   3 │     2  B
   4 │     3  BB

julia> filtrar(dt, "2 == :A" , "'A'== :B")
1×2 DataFrame
 Row │ A    B
     │ Any  Any
─────┼──────────
   1 │ 2    A

julia> contar_renglones(dt, "2 == :A" , "'A'== :B")
1

Reporstat.sumar_columna — Function

sumar_columna(tabla::DataFrame, col::Int)::Number
sumar_columna(tabla::DataFrame, col::String)::Number

Suma todos los valores de una determinada columna en un DataFrame. Para hacer referencia a que columna se desea sumar se pude usar la posición de la columna o el nombre que tiene.

Ejemplo

julia> df = cargar_csv("datos.csv")
4×2 DataFrame
 Row │ x      y
     │ Int64  Int64
─────┼──────────────
   1 │     0     11
   2 │     2     12
   3 │     0     13
   4 │    40     14

julia> sumar_columna(df, 1)
42

julia> sumar_columna(df, "x")
42

Reporstat.sumar_fila — Function

sumar_fila(tabla::DataFrame, fila::Int)::Number

Suma todos los valores de una determinada fila en un DataFrame. La fila se especifica con la posición en la que se encuentra.

Ejemplo

julia> df = cargar_csv("datos.csv")
4×2 DataFrame
 Row │ x      y
     │ Int64  Int64
─────┼──────────────
   1 │     0     11
   2 │     2     12
   3 │     0     13
   4 │    40     14

julia> sumar_fila(df, 2)
14

julia> sumar_fila(df, 4)
54