Covid.jl

Covid.CSV_to_DataFrame — Function

CSV_to_DataFrame(path_url::String, encoding::String="UTF-8")::DataFrame

Lee un archivo CSV con el encoding indicado y regresa un DataFrame.

Ejemplo

julia> df = DataFrameEncode("datos.csv")
julia> df_latin1 = DataFrameEncode("datos.csv", "LATIN1")

Los encodings soportados dependen de la plataforma, obtén la lista de la siguiente manera.

julia> using StringEncodings
julia> encodings()

Covid.data_check — Function

data_check(path_url::String, type::String="PATH", encoding::String="UTF-8")::DataFrame

Crea un DataFrame dado un archivo CSV o una liga al archivo. Se pude especificar el encoding.

Ejemplo

julia> url = "http://www.conapo.gob.mx/work/models/OMI/Datos_Abiertos/DA_IAIM/IAIM_Municipio_2010.csv"
julia> first(data_check(url, "URL", "LATIN1"))
julia> first(data_check("prueba.csv"))

Covid.fechahoy — Method

fechahoy()::String

Crea un string con la fecha de hoy utilizando el formato "yyyymmdd". Año con cuarto dígitos, mes y día con dos.

Ejemplo

julia> fechahoy()
"20210112"

Covid.unzip — Function

unzip(path::String, dest::String="")

Descomprime y guarda el archivo en el destino indicado(dest), si no se proporciona un destino, se guarda en el directorio actual.

Ejemplo

julia> unzip("datos.zip")
julia> unzip("datos.zip", pwd()*"/datos")