Covid.jl

Paquete que tiene como objetivo aprovechar la información a nivel municipal del INEGI, CONAPO y CONEVAL para consolidar en un solo lugar los datos abiertos de la Secretaría de Salud sobre COVID-19 en México, junto con información relevante del municipio de residencia de cada caso.

Indice

Covid.jl

Indice

Funciones

INEGI

Utilidades

Funciones

Covid.CSV_to_DataFrame

Covid.contar_renglones

Covid.data_check

Covid.fechahoy

Covid.filtrar

Covid.jsonparse

Covid.poblacion_entidad

Covid.poblacion_mexico

Covid.poblacion_municipio

Covid.poblacion_todos_entidades

Covid.poblacion_todos_municipios

Covid.sumacolumna

Covid.sumafila

Covid.unzip

INEGI

Covid.poblacion_mexico — Function

poblacion_mexico(token_INEGI::String="")::DataFrame

Regresa un DataFrame con los datos más recientes, a nivel nacional, proporcionados por la API de Indicadores del INEGI. Requiere el token (token_INEGI) de la API, puede obtenerse aquí. Se pude proporcionar el token directamente o por medio de una variable de entorno llamada de la misma manera, token_INEGI.

Ejemplo

julia> ENV["token_INEGI"] = "00000000-0000-0000-0000-000000000000"
"00000000-0000-0000-0000-000000000000"

julia> popu = poblacion_mexico()
1×8 DataFrame
 Row │ lugar   total      hombres    mujeres    porcentaje_hombres  porcentaje_mujeres  porcentaje_indigena  densidad
     │ String  Float64    Float64    Float64    Float64             Float64             Float64              Float64
─────┼────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ México  1.19938e8  5.48552e7  5.74813e7               48.57               51.43              21.4965   60.9642

Covid.poblacion_entidad — Function

poblacion_entidad(cve_entidad::String, token_INEGI::String="")::poblacion

Regresa un una DataFrame con los datos más recientes, por entidad federativa, proporcionados por la API de Indicadores del INEGI. Requiere el token (token_INEGI) de la API, puede obtenerse aquí. Se pude proporcionar el token directamente o por medio de una variable de entorno, de la siguiente manera.

julia> ENV["token_INEGI"] = "00000000-0000-0000-0000-000000000000"

El DataFrame resultante contiene los siguientes datos.

lugar

población total

densidad de población (habitantes por kilómetro cuadrado)

población total hombres

población total mujeres

porcentaje de hombres

porcentaje de mujeres

porcentaje de población que se considera indígena

Note

Área geoestadística estatal (AGEE)

La entidad federativa se codifica de acuerdo con el orden alfabético de sus nombres oficiales, con una longitud de dos dígitos, a partir del 01 en adelante, según el número de entidades federativas que dispongan las leyes vigentes; en este momento son 32 entidades federativas (Aguascalientes 01, Baja California 02,... y Zacatecas 32). Las puedes consultar aquí.

Clave Entidad	Entidad

01	Aguascalientes

02	Baja California

03	Baja California Sur

⋮	⋮

30	Veracruz de Ignacio de la Llave

31	Yucatán

32	Zacatecas

Ejemplo

julia> popu = poblacion_entidad("31", token)
1×8 DataFrame
 Row │ lugar    total      hombres   mujeres   porcentaje_hombres  porcentaje_mujeres  porcentaje_indigena  densidad
     │ String   Float64    Float64   Float64   Float64             Float64             Float64              Float64
─────┼───────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ Yucatán  2.10226e6  963333.0  992244.0             48.9968             51.0032              65.4035   53.0607

Covid.poblacion_municipio — Function

poblacion_municipio(cve_entidad::String, cve_municipio::String, token_INEGI::String="")::DataFrame

Regresa un DataFrame con los datos más recientes, por municipio, proporcionados por la API de Indicadores del INEGI. Requiere el token (token_INEGI) de la API, puede obtenerse aquí. Se pude proporcionar el token directamente o por medio de una variable de entorno llamada de la misma manera, token_INEGI.

Note

Área geoestadística municipal (AGEM)

La clave del municipio está formada por tres números que se asignan de manera ascendente a partir del 001, de acuerdo con el orden alfabético de los nombres de los municipios, aunque a los creados posteriormente a la clavificación inicial, se les asigna la clave geoestadística conforme se vayan creando. Las puedes consultar aquí.

Clave Entidad	Nombre Entidad	Clave Municipio	Nombre Municipio

01	Aguascalientes	001	Aguascalientes

01	Aguascalientes	002	Asientos

01	Aguascalientes	003	Calvillo

⋮	⋮	⋮	⋮

32	Zacatecas	056	Zacatecas

32	Zacatecas	057	Trancoso

32	Zacatecas	058	Santa María de la Paz

Ejemplo

julia> ENV["token_INEGI"] = "00000000-0000-0000-0000-000000000000"
"00000000-0000-0000-0000-000000000000"

julia> popu = poblacion_municipio("01", "002")
1×8 DataFrame
 Row │ lugar                     total    hombres  mujeres  porcentaje_hombres  porcentaje_mujeres  porcentaje_indigena  densidad
     │ String                    Float64  Float64  Float64  Float64             Float64             Float64              Float64
─────┼────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ Aguascalientes, Asientos  45492.0  22512.0  22980.0             48.9519             51.0481              3.63938   84.6325

Covid.poblacion_todos_entidades — Function

poblacion_todos_entidades()::DataFrame

Regresa un DataFrame con los datos poblacionales de todas las entidades.

clave de entidad

nombre oficial de la entidad

población total

densidad de población (habitantes por kilómetro cuadrado)

población total hombres

población total mujeres

porcentaje de hombres

porcentaje de mujeres

porcentaje de población que se considera indígena

Ejemplo

julia> poblacion_todos_entidades()
32×9 DataFrame
 Row │ entidad  entidad_nombre     total         densidad    hombres         mujeres         porcentaje_hombres ⋯
     │ String   String             Float64       Float64     Float64         Float64         Float64            ⋯
─────┼────────────────────────────────────────────────────────────────────────────────────────────────────────── 
   1 │ 01       Aguascalientes      1.185e6     233.729    576638.0        608358.0           48.7672 ⋯
   2 │ 02       Baja California     3.15507e6    46.4066        1.59161e6       1.56346e6     49.7725 
   ⋮ │ ⋮             ⋮               ⋮               ⋮            ⋮           ⋮                 ⋮      
  32 │ 32       Zacatecas           1.49067e6    20.9791   726897.0        763771.0           48.7819

Covid.poblacion_todos_municipios — Function

poblacion_todos_municipios()::DataFrame

Regresa un DataFrame con los datos poblacionales de todos los municipios.

nombre del lugar

clave de entidad

nombre de la entidad

clave de municipio

nombre de municipio

población total

densidad de población (habitantes por kilómetro cuadrado)

población total hombres

población total mujeres

porcentaje de hombres

porcentaje de mujeres

porcentaje de población que se considera indígena

Ejemplo

julia> poblacion_todos_municipios()
2469×11 DataFrame
  Row │ entidad  entidad_nombre  municipio  municipio_nombre              total     densidad   hombres   mujeres   porcentaje_hombres  porcentajes_mujeres ⋯
      │ String   String          String     String                        Float64   Float64    Float64   Float64   Float64             Float64             ⋯
──────┼───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes  001        Aguascalientes                797010.0  744.58     386429.0  410581.0             48.5335              51.4665 ⋯
    2 │ 01       Aguascalientes  002        Asientos                      797010.0  744.58     386429.0  410581.0             48.5335              51.4665 
    ⋮   │    ⋮           ⋮             ⋮                   ⋮                   ⋮          ⋮         ⋮         ⋮              ⋮                    ⋮          ⋱
 2468 │ 32       Zacatecas       057        Trancoso                      138176.0  331.026     66297.0   71879.0             48.482               51.518  ⋯
 2469 │ 32       Zacatecas       058        Santa María de la Paz          16934.0   87.9192     8358.0    8576.0             48.962               51.038  

Utilidades

Covid.data_check — Function

data_check(path_url::String, type::String="PATH", encoding::String="UTF-8")::DataFrame

Crea un DataFrame dado un archivo CSV o una liga al archivo. Se pude especificar el encoding.

Ejemplo

julia> url = "http://www.conapo.gob.mx/work/models/OMI/Datos_Abiertos/DA_IAIM/IAIM_Municipio_2010.csv"
julia> first(data_check(url, "URL", "LATIN1"))
julia> first(data_check("prueba.csv"))

Covid.CSV_to_DataFrame — Function

CSV_to_DataFrame(path_url::String, encoding::String="UTF-8")::DataFrame

Lee un archivo CSV con el encoding indicado y regresa un DataFrame.

Ejemplo

julia> df = DataFrameEncode("datos.csv")
julia> df_latin1 = DataFrameEncode("datos.csv", "LATIN1")

Los encodings soportados dependen de la plataforma, obtén la lista de la siguiente manera.

julia> using StringEncodings
julia> encodings()

Covid.unzip — Function

unzip(path::String, dest::String="")

Descomprime y guarda el archivo en el destino indicado(dest), si no se proporciona un destino, se guarda en el directorio actual.

Ejemplo

julia> unzip("datos.zip")
julia> unzip("datos.zip", pwd()*"/datos")

Covid.jsonparse — Function

jsonparse(url::String)::Dict

Hace un http request al url especificado y convierte el json obtenido del sitio web en un diccionario. En caso de que el servidor devuelva un status distinto a 200, se arroja un error.

Ejemplo

julia> datos = jsonparse("https://sitioweb.com/datos.json")

Covid.fechahoy — Function

fechahoy()::String

Crea un string con la fecha de hoy utilizando el formato "yyyymmdd". Año con cuarto dígitos, mes y día con dos.

Ejemplo

julia> fechahoy()
"20210112"

Covid.filtrar — Function

filtrar(tabla::DataFrame, condiciones...)::DataFrame

Filtra un DataFrame de acuerdo a los parámetros que sean pasados, en este caso los parámetros actúan como expresiones, los nombres de las columnas deben siempre tener el prefijo : , y puede ser especificada por nombre o por numero de columna.

Ejemplo

julia> dt= DataFrame(A = [1,2,2,3],B = ["A","A","B","BB"])
4×2 DataFrame
 Row │ A      B      
     │ Int32  String 
─────┼───────────────
   1 │     1  A
   2 │     2  A
   3 │     2  B
   4 │     3  BB

julia> Filtrar(dt, ":1 == 2")
2×2 DataFrame
 Row │ A    B   
     │ Any  Any 
─────┼──────────
   1 │ 2    A
   2 │ 2    B

En caso de que se trate de una string se deben siempre encerrar entre comillas simples.

julia> Filtrar(dt, "'A' == :2")
2×2 DataFrame
 Row │ A    B   
     │ Any  Any 
─────┼──────────
   1 │ 1    A
   2 │ 2    A

Las condiciones se deben poner en parametros diferentes.

julia> Filtrar(dt, "2 == :A" , "'A'== :B")
1×2 DataFrame
 Row │ A    B   
     │ Any  Any 
─────┼──────────
   1 │ 2    A

Missing docstring.

Missing docstring for select. Check Documenter's build log for details.

Covid.contar_renglones — Function

contar_renglones(tabla::DataFrame, condiciones...)::Number

Llama internamente a la función Covid.filtrar con los mismos argumentos y regresa el numero de renglones que tiene el DataFrame que retorna Covid.filtrar

Ejemplo

julia> dt
4×2 DataFrame
 Row │ A      B      
     │ Int32  String 
─────┼───────────────
   1 │     1  A
   2 │     2  A
   3 │     2  B
   4 │     3  BB

julia> Filtrar(dt, "2 == :A" , "'A'== :B")
1×2 DataFrame
 Row │ A    B   
     │ Any  Any 
─────┼──────────
   1 │ 2    A

julia> contar_renglones(dt, "2 == :A" , "'A'== :B")
1

Covid.sumacolumna — Function

sumacolumna(tabla::DataFrame, col::Int)::Number
sumacolumna(tabla::DataFrame, col::String)::Number

Suma todos los valores de una determinada columna en un DataFrame. Para hacer referencia a que columna se desea sumar se pude usar la posición de la columna o el nombre que tiene.

Ejemplo

julia> df = data_check("datos.csv")
4×2 DataFrame
 Row │ x      y
     │ Int64  Int64
─────┼──────────────
   1 │     0     11
   2 │     2     12
   3 │     0     13
   4 │    40     14

julia> sumacolumna(df, 1)
42

julia> sumacolumna(df, "x")
42

Covid.sumafila — Function

sumafila(tabla::DataFrame, fila::Int)::Number

Suma todos los valores de una determinada fila en un DataFrame. La fila se especifica con la posición en la que se encuentra.

Ejemplo

julia> df = data_check("datos.csv")
4×2 DataFrame
 Row │ x      y
     │ Int64  Int64
─────┼──────────────
   1 │     0     11
   2 │     2     12
   3 │     0     13
   4 │    40     14

julia> sumafila(df, 2)
14

julia> sumafila(df, 4)
54