Reporstat.jl

Paquete que tiene como objetivo aprovechar la información a nivel municipal del INEGI, CONAPO y CONEVAL para consolidar en un solo lugar los datos abiertos de la Secretaría de Salud sobre COVID-19 en México, junto con información relevante del municipio de residencia de cada caso.

Indice

Reporstat.jl

Indice

Funciones

Población

Indicadores de Pobreza

Utilidades

Funciones

Reporstat.cargar_csv

Reporstat.clave

Reporstat.contar_renglones

Reporstat.csv_a_DataFrame

Reporstat.fechahoy

Reporstat.filtrar

Reporstat.idh

Reporstat.indicadores_pobreza

Reporstat.indicadores_pobreza_porcentaje

Reporstat.jsonparse

Reporstat.poblacion_entidad

Reporstat.poblacion_mexico

Reporstat.poblacion_municipio

Reporstat.poblacion_todas_entidades

Reporstat.poblacion_todos_municipios

Reporstat.seleccionar

Reporstat.sumar_columna

Reporstat.sumar_fila

Reporstat.unzip

Población

Datos proporcionados por el INEGI, actualizados al 2020 a excepción del porcentaje de población que se considera indígena, que los últimos datos proporcionados son del 2015 y la extensión territorial que se calcula con la densidad y población total.

Reporstat.idh — Function

idh(cve_entidad::String, cve_municipio::String="")::Number

Regresa el indice de desarrollo humano de una entidad o de un municipio se debe especificar la clave para ambos parametros, si solo se manda el parametro cveentidad_ se regresara el idh de la entidad.Los datos son obtenidos de la pgina oficial de las naciones unidas puedes consultar aquí.

Ejemplo

julia> idh(clave("Campeche"),"002")
0.797
julia> idh(clave("Campeche"),"003")
0.775

Reporstat.poblacion_mexico — Function

poblacion_mexico(token_INEGI::String="")::DataFrame

Regresa un DataFrame con los datos más recientes, a nivel nacional, proporcionados por la API de Indicadores del INEGI. Requiere el token (token_INEGI) de la API, puede obtenerse aquí. Se pude proporcionar el token directamente o por medio de una variable de entorno llamada de la misma manera, token_INEGI.

Ejemplo

julia> ENV["token_INEGI"] = "00000000-0000-0000-0000-000000000000"
"00000000-0000-0000-0000-000000000000"

julia> popu = poblacion_mexico()
1×8 DataFrame
 Row │ lugar   total      hombres    mujeres    porcentaje_hombres  porcentaje_mujeres  porcentaje_indigena  densidad
     │ String  Float64    Float64    Float64    Float64             Float64             Float64              Float64
─────┼────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ México  1.19938e8  5.48552e7  5.74813e7               48.57               51.43              21.4965   60.9642

Reporstat.poblacion_entidad — Function

poblacion_entidad(cve_entidad::String, token_INEGI::String="")::poblacion

Regresa un una DataFrame con los datos más recientes, por entidad federativa, proporcionados por la API de Indicadores del INEGI. Requiere el token (token_INEGI) de la API, puede obtenerse aquí. Se pude proporcionar el token directamente o por medio de una variable de entorno, de la siguiente manera.

julia> ENV["token_INEGI"] = "00000000-0000-0000-0000-000000000000"

El DataFrame resultante contiene los siguientes datos.

lugar

población total

densidad de población (habitantes por kilómetro cuadrado)

población total hombres

población total mujeres

porcentaje de hombres

porcentaje de mujeres

porcentaje de población que se considera indígena

Note

Área geoestadística estatal (AGEE)

La entidad federativa se codifica de acuerdo con el orden alfabético de sus nombres oficiales, con una longitud de dos dígitos, a partir del 01 en adelante, según el número de entidades federativas que dispongan las leyes vigentes; en este momento son 32 entidades federativas (Aguascalientes 01, Baja California 02,... y Zacatecas 32). Las puedes consultar aquí.

Clave Entidad	Entidad

01	Aguascalientes

02	Baja California

03	Baja California Sur

⋮	⋮

30	Veracruz de Ignacio de la Llave

31	Yucatán

32	Zacatecas

Ejemplo

julia> popu = poblacion_entidad("31", token)
1×8 DataFrame
 Row │ lugar    total      hombres   mujeres   porcentaje_hombres  porcentaje_mujeres  porcentaje_indigena  densidad
     │ String   Float64    Float64   Float64   Float64             Float64             Float64              Float64
─────┼───────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ Yucatán  2.10226e6  963333.0  992244.0             48.9968             51.0032              65.4035   53.0607

Reporstat.poblacion_municipio — Function

poblacion_municipio(cve_entidad::String, cve_municipio::String, token_INEGI::String="")::DataFrame

Regresa un DataFrame con los datos más recientes, por municipio, proporcionados por la API de Indicadores del INEGI. Requiere el token (token_INEGI) de la API, puede obtenerse aquí. Se pude proporcionar el token directamente o por medio de una variable de entorno llamada de la misma manera, token_INEGI.

Note

Área geoestadística municipal (AGEM)

La clave del municipio está formada por tres números que se asignan de manera ascendente a partir del 001, de acuerdo con el orden alfabético de los nombres de los municipios, aunque a los creados posteriormente a la clavificación inicial, se les asigna la clave geoestadística conforme se vayan creando. Las puedes consultar aquí.

Clave Entidad	Nombre Entidad	Clave Municipio	Nombre Municipio

01	Aguascalientes	001	Aguascalientes

01	Aguascalientes	002	Asientos

01	Aguascalientes	003	Calvillo

⋮	⋮	⋮	⋮

32	Zacatecas	056	Zacatecas

32	Zacatecas	057	Trancoso

32	Zacatecas	058	Santa María de la Paz

Ejemplo

julia> ENV["token_INEGI"] = "00000000-0000-0000-0000-000000000000"
"00000000-0000-0000-0000-000000000000"

julia> popu = poblacion_municipio("01", "002")
1×8 DataFrame
 Row │ lugar                     total    hombres  mujeres  porcentaje_hombres  porcentaje_mujeres  porcentaje_indigena  densidad
     │ String                    Float64  Float64  Float64  Float64             Float64             Float64              Float64
─────┼────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ Aguascalientes, Asientos  45492.0  22512.0  22980.0             48.9519             51.0481              3.63938   84.6325

Reporstat.poblacion_todas_entidades — Function

poblacion_todas_entidades()::DataFrame

Regresa un DataFrame con los datos poblacionales de todas las entidades.

clave de entidad

nombre oficial de la entidad

población total

densidad de población (habitantes por kilómetro cuadrado)

población total hombres

población total mujeres

porcentaje de hombres

porcentaje de mujeres

porcentaje de población que se considera indígena

Ejemplo

julia> poblacion_todas_entidades()
32×9 DataFrame
 Row │ entidad  entidad_nombre     total         densidad    hombres         mujeres         porcentaje_hombres ⋯
     │ String   String             Float64       Float64     Float64         Float64         Float64            ⋯
─────┼────────────────────────────────────────────────────────────────────────────────────────────────────────── 
   1 │ 01       Aguascalientes      1.185e6     233.729    576638.0        608358.0           48.7672 ⋯
   2 │ 02       Baja California     3.15507e6    46.4066        1.59161e6       1.56346e6     49.7725 
   ⋮ │ ⋮             ⋮               ⋮               ⋮            ⋮           ⋮                 ⋮      
  32 │ 32       Zacatecas           1.49067e6    20.9791   726897.0        763771.0           48.7819

Reporstat.poblacion_todos_municipios — Function

poblacion_todos_municipios()::DataFrame

Regresa un DataFrame con los datos poblacionales de todos los municipios.

nombre del lugar

clave de entidad

nombre de la entidad

clave de municipio

nombre de municipio

población total

densidad de población (habitantes por kilómetro cuadrado)

población total hombres

población total mujeres

porcentaje de hombres

porcentaje de mujeres

porcentaje de población que se considera indígena

Ejemplo

julia> poblacion_todos_municipios()
2469×11 DataFrame
  Row │ entidad  entidad_nombre  municipio  municipio_nombre              total     densidad   hombres   mujeres   porcentaje_hombres  porcentajes_mujeres ⋯
      │ String   String          String     String                        Float64   Float64    Float64   Float64   Float64             Float64             ⋯
──────┼───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes  001        Aguascalientes                797010.0  744.58     386429.0  410581.0             48.5335              51.4665 ⋯
    2 │ 01       Aguascalientes  002        Asientos                      797010.0  744.58     386429.0  410581.0             48.5335              51.4665 
    ⋮   │    ⋮           ⋮             ⋮                   ⋮                   ⋮          ⋮         ⋮         ⋮              ⋮                    ⋮          ⋱
 2468 │ 32       Zacatecas       057        Trancoso                      138176.0  331.026     66297.0   71879.0             48.482               51.518  ⋯
 2469 │ 32       Zacatecas       058        Santa María de la Paz          16934.0   87.9192     8358.0    8576.0             48.962               51.038  

Indicadores de Pobreza

Datos proporcionados por el Consejo Nacional de Evaluación de la Política de Desarrollo Social, actualizados al año 2015.

Diccionario de Datos, Indicadores de pobreza municipal (2015)

Abreviación	Significado

pobreza	Población en situación de pobreza

pobreza_e	Población en situación de pobreza extrema

pobreza_m	Población en situación de pobreza moderada

vul_car	Población vulnerable por carencias

vul_ing	Población vulnerable por ingreso

npnv	Población no pobre y no vulnerable

ic_rezedu	Población con rezago educativo

ic_asalud	Población con carencia por acceso a los servicios de salud

ic_segsoc	Población con carencia por acceso a la seguridad social

ic_cv	Población con carencia por calidad y espacios en la vivienda

ic_sbv	Población con carencia por acceso a los servicios básicos en la vivienda

ic_ali	Población con carencia por acceso a la alimentación

carencias	Población con al menos una carencia social

carencias3	Población con al menos tres carencias sociales

plb	Población con ingreso inferior a la línea de bienestar

plbm	Población con ingreso inferior a la línea de bienestar mínimo

Reporstat.indicadores_pobreza — Function

indicadores_pobreza()::DataFrame

Proporciona el número de personas que cumple con los indicadores de pobreza según el CONEVAL, a nivel municipal.

Los datos son obtenidos de la página oficial de datos abiertos del gobierno federal de México datos.gob.mx Consulta el Diccionario de Datos, Indicadores de pobreza municipal (2015)

julia> df = indicadores_pobreza() 
2457×20 DataFrame
  Row │ entidad  entidad_nombre       municipio  municipio_nombre pobreza  pobreza_e  pobreza_m ⋯
      │ String   String               String     String           Int64    Int64      Int64     ⋯
──────┼────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes       001        Aguascalientes   224949      13650     211299 ⋯
    2 │ 01       Aguascalientes       002        Asientos          25169       2067      23101 ⋯
   ⋮  │ ⋮             ⋮               ⋮               ⋮            ⋮            ⋮          ⋮   

Reporstat.indicadores_pobreza_porcentaje — Function

indicadores_pobreza_porcentaje()::DataFrame

Proporciona el porcentaje de personas que cumple con los indicadores de pobreza según el CONEVAL, a nivel municipal.

Los datos son obtenidos de la página oficial de datos abiertos del gobierno federal de México datos.gob.mx Consulta el Diccionario de Datos, Indicadores de pobreza municipal (2015)

julia> df = indicadores_pobreza_porcentaje() 
2457×20 DataFrame
  Row │ entidad  entidad_nombre       municipio  municipio_nombre pobreza  pobreza_e  pobreza_m ⋯
      │ String   String               String     String           Float64  Float64    Float64   ⋯
──────┼─────────────────────────────────────────────────────────────────────────────────────────
    1 │ 01       Aguascalientes       001        Aguascalientes      26.1        1.6       24.5 ⋯
    2 │ 01       Aguascalientes       002        Asientos            54.0        4.4       49.5 ⋯
   ⋮  │ ⋮             ⋮               ⋮               ⋮            ⋮            ⋮          ⋮    

Utilidades

Utilidades varias que tienen como objetivo facilitar la manipulación de archivos CSV, DataFrames, entre otras.

Reporstat.clave — Function

clave(id::String)::String

Toma como parámetro el nombre de algún municipio o entidad y regresa la clave de este.

Ejemplo

julia> clave("Campeche")
"04"
julia> clave("Calakmul")
"010"

Reporstat.cargar_csv — Function

cargar_csv(path_url::String, type::String="PATH", encoding::String="UTF-8")::DataFrame

Crea un DataFrame dado un archivo CSV o una liga al archivo. Se pude especificar el encoding.

Ejemplo

julia> url = "http://www.conapo.gob.mx/work/models/OMI/Datos_Abiertos/DA_IAIM/IAIM_Municipio_2010.csv"
julia> first(cargar_csv(url, "URL", "LATIN1"))
julia> first(cargar_csv("prueba.csv"))

Reporstat.csv_a_DataFrame — Function

csv_a_DataFrame(path_url::String, encoding::String="UTF-8")::DataFrame

Lee un archivo CSV con el encoding indicado y regresa un DataFrame.

Ejemplo

julia> df = DataFrameEncode("datos.csv")
julia> df_latin1 = DataFrameEncode("datos.csv", "LATIN1")

Los encodings soportados dependen de la plataforma, obtén la lista de la siguiente manera.

julia> using StringEncodings
julia> encodings()

Reporstat.unzip — Function

unzip(path::String, dest::String="")

Descomprime y guarda el archivo en el destino indicado(dest), si no se proporciona un destino, se guarda en el directorio actual.

Ejemplo

julia> unzip("datos.zip")
julia> unzip("datos.zip", pwd()*"/datos")

Reporstat.jsonparse — Function

jsonparse(url::String)::Dict

Hace un http request al url especificado y convierte el json obtenido del sitio web en un diccionario. En caso de que el servidor devuelva un status distinto a 200, se arroja un error.

Ejemplo

julia> datos = jsonparse("https://sitioweb.com/datos.json")

Reporstat.fechahoy — Function

fechahoy()::String

Crea un string con la fecha de hoy utilizando el formato "yyyymmdd". Año con cuarto dígitos, mes y día con dos.

Ejemplo

julia> fechahoy()
"20210112"

Reporstat.filtrar — Function

filtrar(tabla::DataFrame, condiciones...)::DataFrame

Filtra un DataFrame de acuerdo a los parámetros que sean pasados, en este caso los parámetros actúan como expresiones, los nombres de las columnas deben siempre tener el prefijo : , y puede ser especificada por nombre o por numero de columna.

Ejemplo

julia> dt= DataFrame(A = [1,2,2,3],B = ["A","A","B","BB"])
4×2 DataFrame
 Row │ A      B      
     │ Int32  String 
─────┼───────────────
   1 │     1  A
   2 │     2  A
   3 │     2  B
   4 │     3  BB

julia> Filtrar(dt, ":1 == 2")
2×2 DataFrame
 Row │ A    B   
     │ Any  Any 
─────┼──────────
   1 │ 2    A
   2 │ 2    B

En caso de que se trate de una string se deben siempre encerrar entre comillas simples.

julia> Filtrar(dt, "'A' == :2")
2×2 DataFrame
 Row │ A    B   
     │ Any  Any 
─────┼──────────
   1 │ 1    A
   2 │ 2    A

Las condiciones se deben poner en parametros diferentes.

julia> Filtrar(dt, "2 == :A" , "'A'== :B")
1×2 DataFrame
 Row │ A    B   
     │ Any  Any 
─────┼──────────
   1 │ 2    A

Reporstat.seleccionar — Function

seleccionar(Tabla::DataFrame, query::Vector{String})::DataFrame

Selecciona una o varias columnas del DataFrame puede ser por nombre o por número de columna y regresa un nuevo DataFrame con las columnas seleccionadas.

Ejemplo

julia> tabla = DataFrame(A = 1:3, B= 1:3)
3×2 DataFrame
 Row │ A      B
     │ Int32  Int32
─────┼──────────────
   1 │     1      1
   2 │     2      2
   3 │     3      3

julia> q1 = seleccionar(tabla,["1"])
3×1 DataFrame
 Row │ A
     │ Int32
─────┼───────
   1 │     1
   2 │     2
   3 │     3

Reporstat.contar_renglones — Function

contar_renglones(tabla::DataFrame, condiciones...)::Number

Llama internamente a la función Reporstat.filtrar con los mismos argumentos y regresa el numero de renglones que tiene el DataFrame que retorna Reporstat.filtrar

Ejemplo

julia> dt
4×2 DataFrame
 Row │ A      B      
     │ Int32  String 
─────┼───────────────
   1 │     1  A
   2 │     2  A
   3 │     2  B
   4 │     3  BB

julia> Filtrar(dt, "2 == :A" , "'A'== :B")
1×2 DataFrame
 Row │ A    B   
     │ Any  Any 
─────┼──────────
   1 │ 2    A

julia> contar_renglones(dt, "2 == :A" , "'A'== :B")
1

Reporstat.sumar_columna — Function

sumar_columna(tabla::DataFrame, col::Int)::Number
sumar_columna(tabla::DataFrame, col::String)::Number

Suma todos los valores de una determinada columna en un DataFrame. Para hacer referencia a que columna se desea sumar se pude usar la posición de la columna o el nombre que tiene.

Ejemplo

julia> df = cargar_csv("datos.csv")
4×2 DataFrame
 Row │ x      y
     │ Int64  Int64
─────┼──────────────
   1 │     0     11
   2 │     2     12
   3 │     0     13
   4 │    40     14

julia> sumar_columna(df, 1)
42

julia> sumar_columna(df, "x")
42

Reporstat.sumar_fila — Function

sumar_fila(tabla::DataFrame, fila::Int)::Number

Suma todos los valores de una determinada fila en un DataFrame. La fila se especifica con la posición en la que se encuentra.

Ejemplo

julia> df = cargar_csv("datos.csv")
4×2 DataFrame
 Row │ x      y
     │ Int64  Int64
─────┼──────────────
   1 │     0     11
   2 │     2     12
   3 │     0     13
   4 │    40     14

julia> sumar_fila(df, 2)
14

julia> sumar_fila(df, 4)
54